Added a minimum secret length.
Given the way that variationsStream is implemented, the code still calculates all permutations with repetition up to the min length, but these are no longer passed along to be processed and checked against the token. Ideal would be to replace or extend the variationsStream call but that library has not been updated in 9 years.